microservices-version/api-gateway/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
PyJWT==2.8.0
bcrypt==4.1.1
python-dotenv==1.0.0
requests==2.31.0
requests==2.31.0